+++ intergrupo-MergeIFRS9/src/main/java/com/inter/proyecto_intergrupo/repository/parametric/TypificationRepository.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TypificationRepository extends JpaRepository<Typification,Long> {
+public interface TypificationRepository extends JpaRepository<Typification,Integer> {
     List<Typification> findAll();
-    Typification findAllById(Long id);
+    Typification findAllById(int id);
     List<Typification> findByEstado(boolean estado);
     List<Typification> findByAplicaConcil(boolean aplicaConcil);
     List<Typification> findAllByDetalleIgnoreCase(String detalle);
